typecheck adapters (#4220)
* typecheck adapters

* changeset

* fix lockfile

* expose Builder
