+++ .changeset/rich-icons-swim.md
@@ -1,0 +1,5 @@
+---
+'@sveltejs/kit': patch
+---
+
+Expose Server and SSRManifest as public types
