+++ packages/adapter-cloudflare-workers/package.json
@@ -24,6 +24,7 @@
 	"scripts": {
 		"lint": "eslint --ignore-path .gitignore \"**/*.{ts,js,svelte}\" && npm run check-format",
 		"format": "npm run check-format -- --write",
+		"check": "tsc",
 		"check-format": "prettier --check . --config ../../.prettierrc --ignore-path .gitignore"
 	},
 	"dependencies": {
@@ -32,6 +33,6 @@
 	},
 	"devDependencies": {
 		"@cloudflare/kv-asset-handler": "^0.2.0",
-		"@sveltejs/kit": "workspace:*"
+		"typescript": "^4.6.2"
 	}
 }
