+++ packages/adapter-cloudflare/ambient.d.ts
