+++ packages/adapter-netlify/ambient.d.ts
