+++ packages/adapter-netlify/rollup.config.js
@@ -2,24 +2,25 @@ import { nodeResolve } from '@rollup/plugin-node-resolve';
 import commonjs from '@rollup/plugin-commonjs';
 import json from '@rollup/plugin-json';
 
-export default [
-	{
-		input: {
-			handler: 'src/handler.js',
-			shims: 'src/shims.js'
+/** @type {import('rollup').RollupOptions} */
+const config = {
+	input: {
+		handler: 'src/handler.js',
+		shims: 'src/shims.js'
+	},
+	output: [
+		{
+			dir: 'files/cjs',
+			format: 'cjs'
 		},
-		output: [
-			{
-				dir: 'files/cjs',
-				format: 'cjs'
-			},
-			{
-				dir: 'files/esm',
-				format: 'esm'
-			}
-		],
-		plugins: [nodeResolve(), commonjs(), json()],
-		external: (id) => id === '0APP' || id.startsWith('node:'),
-		preserveEntrySignatures: true
-	}
-];
+		{
+			dir: 'files/esm',
+			format: 'esm'
+		}
+	],
+	plugins: [nodeResolve(), commonjs(), json()],
+	external: (id) => id === '0SERVER' || id.startsWith('node:'),
+	preserveEntrySignatures: 'exports-only'
+};
+
+export default config;
