+++ packages/adapter-node/package.json
@@ -25,6 +25,7 @@
 		"dev": "rimraf files && rollup -cw",
 		"build": "rimraf files && rollup -c",
 		"test": "echo \"tests temporarily disabled\" # c8 uvu tests",
+		"check": "tsc",
 		"lint": "eslint --ignore-path .gitignore \"**/*.{ts,js,svelte}\" && npm run check-format",
 		"format": "npm run check-format -- --write",
 		"check-format": "prettier --check . --config ../../.prettierrc --ignore-path .gitignore",
@@ -44,6 +45,7 @@
 		"rimraf": "^3.0.2",
 		"rollup": "^2.60.2",
 		"sirv": "^2.0.0",
+		"typescript": "^4.6.2",
 		"uvu": "^0.5.2"
 	}
 }
