+++ packages/adapter-vercel/package.json
@@ -24,12 +24,14 @@
 	"scripts": {
 		"lint": "eslint --ignore-path .gitignore \"**/*.{ts,js,svelte}\" && npm run check-format",
 		"format": "npm run check-format -- --write",
-		"check-format": "prettier --check . --config ../../.prettierrc --ignore-path .gitignore"
+		"check-format": "prettier --check . --config ../../.prettierrc --ignore-path .gitignore",
+		"check": "tsc"
 	},
 	"dependencies": {
 		"esbuild": "^0.14.21"
 	},
 	"devDependencies": {
-		"@sveltejs/kit": "workspace:*"
+		"@sveltejs/kit": "workspace:*",
+		"typescript": "^4.6.2"
 	}
 }
